+++ CHANGELOG.md
@@ -61,6 +61,11 @@
 
 
 
+* PAPI
+  * Added import of the `certificate` for `akamai_edge_hostname` resource ([I#338](https://github.com/akamai/terraform-provider-akamai/issues/338)).
+  * NOTE: Certificate modification is not allowed.
+
+